CDT DGS sampler in keygen & FFO Babai in NTRUsolve
The Gaussian sampler in keygen is now using no conditional probabilities in the sampler, only a cumulative probability table (and an additional bit for the sign if the value is nonzero), this uses half the number of random bits needed and still gives a smaller Renyi Divergence.

The solve_NTRU_binary_depth0 method now does no Babai reduction (simple rounding over the ring), but uses the FFO Nearest Plane to do Babai reduction with respect to the GSO of the rotations of (f, g) immediately.
Prior to this, we would solve the basis, check if the solution was valid and then do FFO Nearest Plane, but if something goes wrong here this is not detected. Therefore, we replace the simple rounding in depth0 with FFO Nearest Plane, and this improves performance and reduces RAM usage!